MO-1626 RBAC access control - part 1
Ticket: https://dsdmoj.atlassian.net/browse/MO-1626

This is a first part, that introduces some boilerplate code and helpers, and for now will check the roles in order to know which cards to show to the user in the homepage.

The endpoints CRUD operations are not yet protected, that will come next.